rename to `TempVirtualFile` step 1: move to separate file
